fix(android): Se corrige la función #getAmount() y #isPay()
Se cambia el valor de retorno de #getAmount() a Float
Se implementa la función #isPay() para devolver el valor según su transacción

+++ CryptoWalletApp/app/src/main/java/com/cryptowallet/wallet/ITransaction.java
@@ -51,7 +51,7 @@ public interface ITransaction extends Comparable<ITransaction> {
      *
      * @return Cantidad de la transacción.
      */
-    long getAmount();
+    Float getAmount();
 
     /**
      * Obtiene la lista de direcciones que envian alguna cantidad en la transacción.
